Some more checks to data types Set default temperature to 63°C

+++ GTEventGenerator2/Entities/EventEntry.cs
@@ -17,8 +17,8 @@ namespace GTEventGenerator
         public string DriverRegion { get; set; } = "PDI";
 
         #region Skill Properties
-        private int _baseSkill = 80;
-        public int BaseSkill
+        private short _baseSkill = 80;
+        public short BaseSkill
         {
             get => _baseSkill; 
             set
